Refactor: rename voxelization module
