+++ tests/discrete_representation/test_matrix_based_voxelization.py
@@ -7,7 +7,7 @@ import volmdlr
 from volmdlr.core import VolumeModel, BoundingBox
 from volmdlr.primitives3d import Block, Cylinder, Sphere
 from volmdlr.shells import ClosedTriangleShell3D
-from volmdlr.voxelization import MatrixBasedVoxelization
+from volmdlr.discrete_representation import MatrixBasedVoxelization
 
 SHOW_BABYLONJS = False
